SysUI: Simplify Code for Battery Icon Tint
Our customization allows optional tinting of battery icons. Due to recent changes in battery theming, rendering colors on vector drawables was useless as the default tint was set to a dark color
lets fix this by simplifying the code and allowing the entire drawable to be tinted.
Fix a major issue where battery level text and background (not battery fill) was not tinting in a light statusbar background,call dontstress to update battery tint immediately.
All in all, maintain integrity with themes.

Signed-off-by: Varun Date <date.varun123@gmail.com>