Only show keyguard panel if on lockscreen + no activity on top
There is a bug in window manager, where if the keyboard is collapsed,
the reference to the window is lost when destroying its surface. This in
turn prevents the ime window from being reordered later on. For normal apps
this is not an issue, but since we're above normal apps (KEYGUARD_PANEL),
this rises an issue for us in this scenario :

1) Set password lockscreen
2) Swipe up to unlock
3) Collapse keyboard
4) Launch emergency dialer (SHOW_WHEN_LOCKED)
5) Press home button

Notice that all touch events are going to the external keyguard.
If the keyboard is not collapsed in step 3, everything works and the reason
is the reference to the input method window in window manager isn't lost
and it gets re-ordered behind the windows.

To workaround this issue, avoid removing/re-adding the window and instead
rely on window manager to control its visibility.

Change-Id: I59dd41547902ae5e23bbeb0a01ee27f11fc29dfc